+++ examples/sql/mysql.propertystorage.sql
@@ -1,0 +1,7 @@
+CREATE TABLE propertystorage (
+    id INT UNSIGNED NOT NULL PRIMARY KEY AUTO_INCREMENT,
+    path VARBINARY(1024) NOT NULL,
+    name VARBINARY(100) NOT NULL,
+    value MEDIUMBLOB
+);
+CREATE UNIQUE INDEX path_property ON propertystorage (path(600), name(100));