+++ tests/Sabre/DAV/Browser/PluginTest.php
@@ -9,10 +9,12 @@ require_once 'Sabre/DAV/AbstractServer.php';
 
 class PluginTest extends DAV\AbstractServer{
 
+    protected $plugin;
+
     function setUp() {
 
         parent::setUp();
-        $this->server->addPlugin(new Plugin());
+        $this->server->addPlugin($this->plugin = new Plugin());
 
     }
 
@@ -27,7 +29,7 @@ class PluginTest extends DAV\AbstractServer{
         $this->server->httpRequest = ($request);
         $this->server->exec();
 
-        $this->assertEquals(200, $this->response->status);
+        $this->assertEquals(200, $this->response->status, "Incorrect status received. Full response body: " . $this->response->getBodyAsString());
         $this->assertEquals(array(
             'Content-Type' => 'text/html; charset=utf-8',
             'Content-Security-Policy' => "img-src 'self'; style-src 'unsafe-inline';"
@@ -35,15 +37,14 @@ class PluginTest extends DAV\AbstractServer{
             $this->response->headers
         );
 
-        $this->assertTrue(strpos($this->response->body, 'Index for dir/') !== false);
-        $this->assertTrue(strpos($this->response->body, '<a href="/dir/child.txt"><img src="/?sabreAction=asset&assetName=icons%2Ffile.png" alt="" width="24" />')!==false);
+        $this->assertTrue(strpos($this->response->body, '<title>dir/') !== false);
+        $this->assertTrue(strpos($this->response->body, '<a href="/dir/child.txt">')!==false);
 
     }
-
-    function testNotFound() {
+    function testCollectionGetRoot() {
 
         $serverVars = array(
-            'REQUEST_URI'    => '/random',
+            'REQUEST_URI'    => '/',
             'REQUEST_METHOD' => 'GET',
         );
 
@@ -51,7 +52,27 @@ class PluginTest extends DAV\AbstractServer{
         $this->server->httpRequest = ($request);
         $this->server->exec();
 
-        $this->assertEquals(404, $this->response->status);
+        $this->assertEquals(200, $this->response->status, "Incorrect status received. Full response body: " . $this->response->getBodyAsString());
+        $this->assertEquals(array(
+            'Content-Type' => 'text/html; charset=utf-8',
+            'Content-Security-Policy' => "img-src 'self'; style-src 'unsafe-inline';"
+            ),
+            $this->response->headers
+        );
+
+        $this->assertTrue(strpos($this->response->body, '<title>/') !== false);
+        $this->assertTrue(strpos($this->response->body, '<a href="/dir/">')!==false);
+        $this->assertTrue(strpos($this->response->body, '<span class="btn disabled">')!==false);
+
+    }
+
+    function testGETPassthru() {
+
+        $request = new HTTP\Request('GET', '/random');
+        $response = new HTTP\Response();
+        $this->assertNull(
+            $this->plugin->httpGet($request, $response)
+        );
 
     }
 
@@ -113,4 +134,28 @@ class PluginTest extends DAV\AbstractServer{
 
     }
 
+    function testGetAsset() {
+
+        $request = new HTTP\Request('GET', '/?sabreAction=asset&assetName=favicon.ico');
+        $this->server->httpRequest = $request;
+        $this->server->exec();
+
+        $this->assertEquals(200, $this->response->getStatus(), 'Error: ' . $this->response->body);
+        $this->assertEquals([
+            'Content-Type' => 'image/vnd.microsoft.icon',
+            'Content-Length' => '4286',
+            'Cache-Control' => 'public, max-age=1209600',
+        ], $this->response->getHeaders());
+
+    }
+
+    function testGetAsset404() {
+
+        $request = new HTTP\Request('GET', '/?sabreAction=asset&assetName=flavicon.ico');
+        $this->server->httpRequest = $request;
+        $this->server->exec();
+
+        $this->assertEquals(404, $this->response->getStatus(), 'Error: ' . $this->response->body);
+
+    }
 }