+++ tests/Sabre/DAV/Property/SupportedReportSetTest.php
@@ -64,7 +64,12 @@ class SupportedReportSetTest extends DAV\AbstractServer {
     function testCustomReport() {
 
         // Intercepting the report property
-        $this->server->on('afterGetProperties', [$this,'addProp']);
+        $this->server->on('propFind', function(DAV\PropFind $propFind, DAV\INode $node) {
+            if ($prop = $propFind->get('{DAV:}supported-report-set')) {
+                $prop->addReport('{http://www.rooftopsolutions.nl/testnamespace}myreport');
+                $prop->addReport('{DAV:}anotherreport');
+            }
+        },200);
 
         $xml = '<?xml version="1.0"?>
 <d:propfind xmlns:d="DAV:">
@@ -107,20 +112,5 @@ class SupportedReportSetTest extends DAV\AbstractServer {
 
     }
 
-    /**
-     * This method is used as a callback for afterGetProperties
-     */
-    function addProp($path, &$properties) {
-
-        if (isset($properties[200]['{DAV:}supported-report-set'])) {
-            $properties[200]['{DAV:}supported-report-set']->addReport('{http://www.rooftopsolutions.nl/testnamespace}myreport');
-            $properties[200]['{DAV:}supported-report-set']->addReport('{DAV:}anotherreport');
-        }
-
-    }
-
-
-
 }
 
-?>